preparado para o sistema piloto

+++ database/migrations/202504131259403-create-saques.js
@@ -13,7 +13,7 @@ export default {
         allowNull: false,
       },
       valor: {
-        type: DataTypes.FLOAT,
+        type: DataTypes.BIGINT,
         allowNull: false,
       },
       taxa: {
