+++ src/models/Emprestimo.ts
@@ -25,7 +25,7 @@ export default class Emprestimo extends Model {
   @Column(DataType.INTEGER)
   id!: number;
 
-  @Column({ type: DataType.FLOAT, allowNull: false })
+  @Column({ type: DataType.BIGINT, allowNull: false })
   valor!: number;
 
   @Column({ type: DataType.INTEGER, allowNull: false })
